requirements.txt
torch
torchinfo
wandb
koila
pytorch-lightning
self-attention-cv